# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-shop-scout",
    "version": "0.1.0",
    "description": "Live marketplace intelligence over eBay Browse — listings search and is-this-a-deal verdicts\n\nEvery paid route returns the purchased artifact in the 200 response body. Pay in USDC on Base or Solana — your client picks the rail.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com",
      "url": "https://github.com/nirholas/x402-shop-scout"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4022",
      "description": "Local development"
    },
    {
      "url": "https://your-deployment.example.com",
      "description": "Your deployment"
    }
  ],
  "tags": [
    {
      "name": "paid",
      "description": "x402-paid routes. An unpaid request answers 402 with both rails."
    },
    {
      "name": "free",
      "description": "Free routes. No payment required."
    }
  ],
  "paths": {
    "/search": {
      "get": {
        "operationId": "searchListings",
        "summary": "Search live marketplace listings with prices, shipping, condition and seller reputation.",
        "description": "Full-text search across the marketplace. Each listing is normalized to a stable shape: item id, title, price and currency, condition, seller username with feedback percentage and score, buying option (fixed price / auction / best offer), shipping cost and type, location, and the listing URL. Shipping is returned separately from price so a caller can compute landed cost.\n\nPrice: **$0.002** in USDC — payable on Base (EVM) or Solana.",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search term, e.g. `sony wh-1000xm5`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum listings to return (1–50).",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "normalized listings with price, shipping cost, condition, seller feedback, location and URL",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt (rail, network, transaction, payer).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "query": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "listings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "itemId": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "price": {
                            "type": "object",
                            "properties": {
                              "value": {
                                "type": "integer"
                              },
                              "currency": {
                                "type": "string"
                              }
                            }
                          },
                          "condition": {
                            "type": "string"
                          },
                          "seller": {
                            "type": "object",
                            "properties": {
                              "username": {
                                "type": "string"
                              },
                              "feedbackPercentage": {
                                "type": "number"
                              },
                              "feedbackScore": {
                                "type": "integer"
                              }
                            }
                          },
                          "buyingOption": {
                            "type": "string"
                          },
                          "shipping": {
                            "type": "object",
                            "properties": {
                              "cost": {
                                "type": "number"
                              },
                              "currency": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              }
                            }
                          },
                          "location": {
                            "type": "string"
                          },
                          "itemWebUrl": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": [
                              "null",
                              "string"
                            ]
                          }
                        }
                      }
                    },
                    "payment": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "rail": {
                          "type": "string"
                        },
                        "network": {
                          "type": "string"
                        },
                        "transaction": {
                          "type": "string"
                        },
                        "payer": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "string"
                        },
                        "asset": {
                          "type": "string"
                        },
                        "resource": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "source": "fixture",
                  "query": "sony wh-1000xm5",
                  "count": 2,
                  "listings": [
                    {
                      "itemId": "v1|120003|0",
                      "title": "Sony WH1000XM5 - broken headband, works with tape",
                      "price": {
                        "value": 119,
                        "currency": "USD"
                      },
                      "condition": "For parts or not working",
                      "seller": {
                        "username": "parts_bin_pete",
                        "feedbackPercentage": 94.9,
                        "feedbackScore": 812
                      },
                      "buyingOption": "AUCTION",
                      "shipping": {
                        "cost": 12.4,
                        "currency": "USD",
                        "type": "STANDARD"
                      },
                      "location": "Tulsa, OK",
                      "itemWebUrl": "https://www.ebay.com/itm/120003",
                      "imageUrl": null
                    },
                    {
                      "itemId": "v1|120002|0",
                      "title": "Sony WH-1000XM5 Headphones Silver - Open Box, complete",
                      "price": {
                        "value": 219.99,
                        "currency": "USD"
                      },
                      "condition": "Open box",
                      "seller": {
                        "username": "bigbox_returns",
                        "feedbackPercentage": 98.2,
                        "feedbackScore": 31005
                      },
                      "buyingOption": "FIXED_PRICE",
                      "shipping": {
                        "cost": 0,
                        "currency": "USD",
                        "type": "FREE"
                      },
                      "location": "Reno, NV",
                      "itemWebUrl": "https://www.ebay.com/itm/120002",
                      "imageUrl": null
                    }
                  ],
                  "payment": {
                    "success": true,
                    "rail": "evm",
                    "network": "base-sepolia",
                    "transaction": "0x9c1f…",
                    "payer": "0xA11ce…",
                    "amount": "2000",
                    "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
                    "resource": "http://localhost:4022/search"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
    "/deal-check/{itemId}": {
      "get": {
        "operationId": "dealCheck",
        "summary": "Grade one listing against comparable listings and return a deal verdict with the arithmetic.",
        "description": "Fetches the listing, pulls comparable active listings, and compares **total cost** — price plus shipping — against their median. Returns one of `great-deal` (≥20% below median), `good-deal` (8–20% below), `fair-price` (within ±10%), `overpriced` (≥10% above), or `insufficient-data` (fewer than two comparables). The comparable sample used is included, so the verdict is auditable rather than asserted.\n\neBay item ids contain `|` characters — URL-encode the id (`v1%7C120002%7C0`) or pass it raw; both work.\n\nPrice: **$0.002** in USDC — payable on Base (EVM) or Solana.",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "description": "Marketplace item id, e.g. `v1|120002|0` (URL-encoding optional).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a verdict (`great-deal` … `overpriced`), the comparable sample, median/mean total cost, delta, percentage, and plain-English reasoning",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt (rail, network, transaction, payer).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "checkedAt": {
                      "type": "string"
                    },
                    "item": {
                      "type": "object",
                      "properties": {
                        "itemId": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "price": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "number"
                            },
                            "currency": {
                              "type": "string"
                            }
                          }
                        },
                        "condition": {
                          "type": "string"
                        },
                        "seller": {
                          "type": "object",
                          "properties": {
                            "username": {
                              "type": "string"
                            },
                            "feedbackPercentage": {
                              "type": "number"
                            },
                            "feedbackScore": {
                              "type": "integer"
                            }
                          }
                        },
                        "buyingOption": {
                          "type": "string"
                        },
                        "shipping": {
                          "type": "object",
                          "properties": {
                            "cost": {
                              "type": "integer"
                            },
                            "currency": {
                              "type": "string"
                            },
                            "type": {
                              "type": "string"
                            }
                          }
                        },
                        "location": {
                          "type": "string"
                        },
                        "itemWebUrl": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      }
                    },
                    "totalCost": {
                      "type": "number"
                    },
                    "comparables": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer"
                        },
                        "priceRange": {
                          "type": "object",
                          "properties": {
                            "min": {
                              "type": "number"
                            },
                            "max": {
                              "type": "integer"
                            }
                          }
                        },
                        "median": {
                          "type": "number"
                        },
                        "mean": {
                          "type": "number"
                        },
                        "sample": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "itemId": {
                                "type": "string"
                              },
                              "title": {
                                "type": "string"
                              },
                              "totalCost": {
                                "type": "integer"
                              },
                              "condition": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "verdict": {
                      "type": "string"
                    },
                    "deltaVsMedian": {
                      "type": "number"
                    },
                    "percentVsMedian": {
                      "type": "integer"
                    },
                    "reasoning": {
                      "type": "string"
                    },
                    "payment": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "rail": {
                          "type": "string"
                        },
                        "network": {
                          "type": "string"
                        },
                        "transaction": {
                          "type": "string"
                        },
                        "payer": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "string"
                        },
                        "asset": {
                          "type": "string"
                        },
                        "resource": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "source": "fixture",
                  "checkedAt": "2026-08-07T12:00:00.000Z",
                  "item": {
                    "itemId": "v1|120002|0",
                    "title": "Sony WH-1000XM5 Headphones Silver - Open Box, complete",
                    "price": {
                      "value": 219.99,
                      "currency": "USD"
                    },
                    "condition": "Open box",
                    "seller": {
                      "username": "bigbox_returns",
                      "feedbackPercentage": 98.2,
                      "feedbackScore": 31005
                    },
                    "buyingOption": "FIXED_PRICE",
                    "shipping": {
                      "cost": 0,
                      "currency": "USD",
                      "type": "FREE"
                    },
                    "location": "Reno, NV",
                    "itemWebUrl": "https://www.ebay.com/itm/120002",
                    "imageUrl": null
                  },
                  "totalCost": 219.99,
                  "comparables": {
                    "count": 2,
                    "priceRange": {
                      "min": 131.4,
                      "max": 248
                    },
                    "median": 189.7,
                    "mean": 189.7,
                    "sample": [
                      {
                        "itemId": "v1|120001|0",
                        "title": "Sony WH-1000XM5 Wireless Noise Canceling Headphones - Black",
                        "totalCost": 248,
                        "condition": "New"
                      },
                      {
                        "itemId": "v1|120003|0",
                        "title": "Sony WH1000XM5 - broken headband, works with tape",
                        "totalCost": 131.4,
                        "condition": "For parts or not working"
                      }
                    ]
                  },
                  "verdict": "overpriced",
                  "deltaVsMedian": 30.29,
                  "percentVsMedian": 16,
                  "reasoning": "Item total cost (price + shipping) is $219.99 vs a median of $189.70 across 2 comparable active listings (+16%). Above market — comparable listings are cheaper.",
                  "payment": {
                    "success": true,
                    "rail": "solana",
                    "network": "solana",
                    "transaction": "5xkQ…",
                    "payer": "9wFh…",
                    "amount": "2000",
                    "asset": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
                    "resource": "http://localhost:4022/deal-check/v1|120002|0"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Liveness probe (free)",
        "description": "Free. Reports the active data source and the payment rails this instance advertises.",
        "tags": [
          "free"
        ],
        "responses": {
          "200": {
            "description": "Service status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "service": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    },
                    "rails": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rail": {
                            "type": "string"
                          },
                          "network": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/.well-known/x402": {
      "get": {
        "operationId": "manifest",
        "summary": "x402 discovery manifest (free)",
        "description": "Free. Machine-readable description of every paid resource and both payment rails.",
        "tags": [
          "free"
        ],
        "responses": {
          "200": {
            "description": "Discovery manifest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentRequirements": {
        "type": "object",
        "description": "One accepted way to pay, as defined by the x402 protocol.",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "description",
          "mimeType",
          "payTo",
          "maxTimeoutSeconds",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "enum": [
              "exact"
            ]
          },
          "network": {
            "type": "string",
            "enum": [
              "base-sepolia",
              "base",
              "solana",
              "solana-devnet"
            ],
            "description": "The settlement network. Both an EVM and a Solana entry are always offered."
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "Price in USDC atomic units (6 decimals). $0.002 → \"2000\"."
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string",
            "description": "EVM address or Solana base58 address."
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string",
            "description": "USDC contract address (EVM) or SPL mint (Solana)."
          },
          "extra": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "PaymentRequiredResponse": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. The body lists both rails — pay with USDC on Base or on Solana, then retry with an `X-PAYMENT` header.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PaymentRequiredResponse"
            },
            "example": {
              "x402Version": 1,
              "error": "X-PAYMENT header is required",
              "accepts": [
                {
                  "scheme": "exact",
                  "network": "base-sepolia",
                  "maxAmountRequired": "2000",
                  "resource": "http://localhost:4022/search",
                  "description": "Search live marketplace listings with prices, shipping, condition and seller reputation.",
                  "mimeType": "application/json",
                  "payTo": "0x40252CFDF8B20Ed757D61ff157719F33Ec332402",
                  "maxTimeoutSeconds": 60,
                  "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
                  "extra": {
                    "name": "USDC",
                    "version": "2"
                  }
                },
                {
                  "scheme": "exact",
                  "network": "solana",
                  "maxAmountRequired": "2000",
                  "resource": "http://localhost:4022/search",
                  "description": "Search live marketplace listings with prices, shipping, condition and seller reputation.",
                  "mimeType": "application/json",
                  "payTo": "WwwuGbqHrwF5RG89KhUbmRWEvjnRH9k5kVM5p7T3WwW",
                  "maxTimeoutSeconds": 60,
                  "asset": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
                  "extra": {
                    "name": "USD Coin",
                    "decimals": 6,
                    "feePayer": "2wKupLR9q6wXYppw8Gr2NvWxKBUqm4PPJKkQfoxHDBg4",
                    "amount": "2000"
                  }
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed request parameters.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UpstreamError": {
        "description": "The upstream provider failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}